+++ CHANGELOG.md
@@ -2,6 +2,64 @@
 
 All notable changes to the **FlexSeedScheme** (FSS) package are documented here.
 
+## 1.1.0
+
+**Nov 17, 2022**
+
+**NEW**
+
+* Added new `FlexTones` method `noOnMainsTint`, that can return a new instance of its configurations with tone mapping for is main **on* colors set to tone, 0 (black) or 100 (white), depending on what is appropriate for its main color tone. Main colors are primary, secondary, tertiary, error and their containers. The method works on any configured `FlexTones`, also custom ones, not only the built-in ones.
+
+* Added new `FlexTones` method `noOnSurfacesTint`, that can return a new instance of its configurations with tone mapping for is surface **on* colors set to tone, 0 (black) or 100 (white), depending on what is appropriate for its surface color tone. Surface colors are background, surface, surfaceVariant and inverseSurface. The method works on any configured `FlexTones`, also custom ones, not only the built-in ones.
+
+* To be able to support creating single hue seeded ColorScheme, the `FlexCorePalette.fromSeeds` got property
+  `tertiaryHueRotation`, that controls the hue degree rotation from primary key color, when tertiary key color is not provided. The `tertiaryHueRotation` defaults 60 degrees, same as previously from Material 3 color system hard coded value.
+
+* Added two new `FlexTones`. 
+  - `FlexTones.oneHue` that set `tertiaryHueRotation` to 0, so we can create a mono hue palette if we only provide primary key color as seed.
+  - `FlexTones.vividBackground` that is a copy of `FlexTones.vividSurfaces` but with tone mapping for `background` and `surface` swapped.
+
+**CHANGE**
+
+Tone mappings for several example/pre-configured `FlexTones` were slightly modified to provide improved output. Most significantly mappings for `FlexTones.vividSurfaces` were modified to provide a more usable and improved vivid tinted surfaces, while still offering a slightly more tinted surface design than `FlexTones.vividSurfaces`. Change details are as follows:
+
+* The `FlexTones.vivid`, brightness **light**:
+  - Tone `surfaceTone` was changed from 99 to 98.
+* The `FlexTones.vivid`, brightness **dark**:
+  - Tone `onErrorContainerTone` was changed from 90 to 80 (dark M3 default).
+  - Tone `backgroundTone` was changed from 10 to 5.
+* The `FlexTones.vividSurfaces` brightness **light**:
+  - Tone `onPrimaryTone` was changed from 95 to 98.
+  - Tone `onSecondaryTone` was changed from 95 to 98.
+  - Tone `onTertiaryTone` was changed from 95 to 98.
+  - Tone `onErrorTone` was changed from 95 to 98.
+  - Tone `primaryContainerTone` was changed from 80 to 90 (light M3 default).
+  - Tone `secondaryContainerTone` was changed from 80 to 90 (light M3 default).
+  - Tone `tertiaryContainerTone` was changed from 80 to 90 (light M3 default).
+  - Tone `errorContainerTone` was changed from 80 to 90 (light M3 default).
+  - Tone `surfaceVariantTone` was changed from 80 to 90 (light M3 default).
+  - Tone `backgroundTone` was changed from 90 to 98.
+  - `neutralChroma` was changed from 8 to 6.
+  - `neutralVariantChroma` was changed from 16 to 12.
+* The `FlexTones.vividSurfaces` brightness **dark**:
+  - Tone `primaryContainerTone` was changed from 40 to 20.
+  - Tone `tertiaryContainerTone` was changed from 40 to 30 (dark M3 default).
+  - Tone `primaryContainerTone` was changed from 40 to 30 (dark M3 default).
+  - Tone `onErrorTone` was changed from 20 to 30 (dark M3 default).
+  - Tone `onErrorContainerTone` was changed from 90 to 80 (dark M3 default).
+  - Tone `surfaceTone` was changed from 10 to 20.
+  - Tone `surfaceVariantTone` was changed from 40 to 30 (dark M3 default).
+  - Tone `backgroundTone` was changed from 20 to 10.
+  - Tone `onSurfaceVariantTone` was changed from 90 to 95.
+  - Tone `onInverseSurfaceTone` was changed from 30 to 20 (dark M3 default).
+  - `neutralChroma` was changed from 8 to 6.
+  - `neutralVariantChroma` was changed from 16 to 12.
+* The `FlexTones.ultraContrast` brightness **light**: 
+  - Tone `primaryTone` was changed from 30 to 20.
+* The `FlexTones.ultraContrast` brightness **dark**:
+  - Tone `surfaceTone` was changed from 10 to 5.
+  - Tone `backgroundTone` was changed from 10 to 5.
+
 ## 1.0.1
 
 **Sep 2, 2022**